Tokenizer: separation of duty.
