+++ NTDLS.Katzebase.Engine/Parsers/Tokens/Tokenizer.TryCompareNext.cs
@@ -1,0 +1,42 @@
+﻿namespace NTDLS.Katzebase.Engine.Parsers.Tokens
+{
+    internal partial class Tokenizer
+    {
+        /// <summary>
+        /// Returns true if the next token is in the given array, using the given delimiters.
+        /// </summary>
+        public bool TryCompareNext(TryNextTokenProc comparer, char[] delimiters)
+        {
+            var token = GetNext(delimiters);
+            return comparer(token);
+        }
+
+        /// <summary>
+        /// Returns true if the next token is in the given array, using the standard delimiters.
+        /// </summary>
+        public bool TryCompareNext(TryNextTokenProc comparer)
+        {
+            var token = GetNext();
+            return comparer(token);
+        }
+
+        /// <summary>
+        /// Returns true if the next token is in the given array, using the given delimiters.
+        /// Regardless of whether a match was made, the token which was parsed it returned via outFoundToken.
+        /// </summary>
+        public bool TryCompareNext(TryNextTokenComparerProc comparer, string[] givenTokens, char[] delimiters, out string outFoundToken)
+        {
+            outFoundToken = GetNext(delimiters);
+
+            foreach (var givenToken in givenTokens)
+            {
+                if (comparer(outFoundToken, givenToken))
+                {
+                    return true;
+                }
+            }
+
+            return false;
+        }
+    }
+}
