+++ NTDLS.Katzebase.Engine/Parsers/Tokens/Tokenizer.Whitespace.cs
@@ -1,0 +1,29 @@
+﻿namespace NTDLS.Katzebase.Engine.Parsers.Tokens
+{
+    internal partial class Tokenizer
+    {
+        /// <summary>
+        /// Moves the caret past any whitespace, does not record breadcrumb.
+        /// </summary>
+        private void InternalEatWhiteSpace()
+        {
+            while (_caret < _text.Length && char.IsWhiteSpace(_text[_caret]))
+            {
+                _caret++;
+            }
+        }
+
+        /// <summary>
+        /// Moves the caret past any whitespace.
+        /// </summary>
+        public void EatWhiteSpace()
+        {
+            RecordBreadcrumb();
+
+            while (_caret < _text.Length && char.IsWhiteSpace(_text[_caret]))
+            {
+                _caret++;
+            }
+        }
+    }
+}
